Add togglable score / player team details
